Restructure part 3 - Move providers to pages they're used - Rename to .service.ts - Fix import paths - Remove unused services and components (TaskSelect component and QuestionService)
