+++ src/app/pages/home/home.ts
@@ -8,11 +8,11 @@ import {
 } from 'ionic-angular'
 
 import { DefaultTask } from '../../../assets/data/defaultConfig'
-import { HomeController } from '../../providers/home-controller'
-import { KafkaService } from '../../providers/kafka-service'
-import { NotificationService } from '../../providers/notification-service'
-import { SchedulingService } from '../../providers/scheduling-service'
-import { StorageService } from '../../providers/storage-service'
+import { HomeController } from '../../core/services/home-controller.service'
+import { KafkaService } from '../../core/services/kafka.service'
+import { NotificationService } from '../../core/services/notification.service'
+import { SchedulingService } from '../../core/services/scheduling.service'
+import { StorageService } from '../../core/services/storage.service'
 import { LocKeys } from '../../shared/enums/localisations'
 import { StorageKeys } from '../../shared/enums/storage'
 import { Task, TasksProgress } from '../../shared/models/task'
