+++ src/app/pages/questions/questions.ts
@@ -7,12 +7,12 @@ import {
   ViewController
 } from 'ionic-angular'
 
-import { AnswerService } from '../../providers/answer-service'
-import { TimeStampService } from '../../providers/timestamp-service'
 import { LocKeys } from '../../shared/enums/localisations'
 import { Question, QuestionType } from '../../shared/models/question'
 import { TranslatePipe } from '../../shared/pipes/translate/translate'
 import { FinishPage } from '../finish/finish'
+import { AnswerService } from './services/answer.service'
+import { TimeStampService } from './services/timestamp.service'
 
 @Component({
   selector: 'page-questions',
@@ -186,7 +186,9 @@ export class QuestionsPage {
   navigateToFinishPage() {
     this.navCtrl.push(FinishPage, {
       endText: this.navParams.data.endText,
-      associatedTask: this.navParams.data.associatedTask
+      associatedTask: this.navParams.data.associatedTask,
+      answers: this.answerService.answers,
+      timestamps: this.timestampService.timestamps
     })
   }
