+++ src/app/pages/questions/services/answer.service.ts
@@ -3,7 +3,7 @@ import 'rxjs/add/operator/map'
 import { HttpClient } from '@angular/common/http'
 import { Injectable } from '@angular/core'
 
-import { Answer } from '../shared/models/answer'
+import { Answer } from '../../../shared/models/answer'
 
 @Injectable()
 export class AnswerService {
