+++ src/app/pages/settings/settings.ts
@@ -14,10 +14,10 @@ import {
   DefaultSettingsWeeklyReport,
   LanguageMap
 } from '../../../assets/data/defaultConfig'
-import { ConfigService } from '../../providers/config-service'
-import { HomeController } from '../../providers/home-controller'
-import { SchedulingService } from '../../providers/scheduling-service'
-import { StorageService } from '../../providers/storage-service'
+import { ConfigService } from '../../core/services/config.service'
+import { HomeController } from '../../core/services/home-controller.service'
+import { SchedulingService } from '../../core/services/scheduling.service'
+import { StorageService } from '../../core/services/storage.service'
 import { LocKeys } from '../../shared/enums/localisations'
 import { StorageKeys } from '../../shared/enums/storage'
 import {
