--- src/app/pages/taskselect/taskselect.scss
@@ -1,27 +1,0 @@
-page-taskselect {
-  .toolbar-background {
-    background-color: $cl-secondary;
-  }
-
-  .toolbar-title {
-    color: $cl-font;
-  }
-
-  .bar-button {
-    color: $cl-font;
-  }
-
-  .item {
-    margin-bottom: $item-separation-distance;
-    background-color: $cl-primary-40 !important;
-  }
-
-  .item:last-child {
-    margin-bottom: 0;
-  }
-
-  .item:hover {
-    background-color: $cl-primary !important;
-    cursor: pointer;
-  }
-}
